Use int32_t for kThreadGroupSize in warp_per_row (#1332)
Summary:
Pull Request resolved: https://github.com/pytorch/FBGEMM/pull/1332

Use int32_t for kThreadGroupSize instead of size_t.  Using size_t
causes a slowdown in some cases.

Reviewed By: jianyuh

Differential Revision: D39568569

fbshipit-source-id: 7a1e56f96a0e9929fb9600156f22fe3c58a3c59e

+++ fbgemm_gpu/codegen/embedding_backward_split_template.cu
@@ -518,7 +518,7 @@ template <
     typename grad_t,
     typename cache_t,
     size_t kMaxVecsPerThread,
-    size_t kThreadGroupSize = kWarpSize>
+    int32_t kThreadGroupSize = kWarpSize>
 __global__
 __launch_bounds__(kBackwardMaxThreads)
 void
